pipeline/requirements.txt
easyocr>=1.7.0
paddleocr>=3.0.0
deep-translator>=1.11.0
Pillow>=9.5.0
numpy>=1.26.0
opencv-python-headless>=4.9.0
huggingface_hub>=0.30.0
onnxruntime>=1.24.0
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0
transformers>=4.35.0
sentencepiece>=0.1.99
simple-lama-inpainting>=0.1.0
